plot.glmertree2 {glmertree}	R Documentation

Plotting (Generalized) Linear Mixed Model Trees

Description

plot method for (g)lmertree objects.

Usage

plot.lmertree2(x, which = "all", ask = TRUE, type = "extended", 
    observed = TRUE, fitmean = TRUE, fit.ranef = "constant", fit.fixef = "constant", 
    tp_args = list(), drop_terminal = TRUE, ...)

plot.glmertree2(x, which = "all", ask = TRUE, type = "extended", 
    observed = TRUE, fitmean = TRUE, fit.ranef = "constant", fit.fixef = "constant",
    tp_args = list(), drop_terminal = TRUE, ...)

Arguments

x
an object of class lmertree or glmertree.

which
character; "all" (default), "tree" or "random". Specifies whether, tree, random effects or both should be plotted.

ask
logical. Should user be asked for input, before a new figure is drawn?

type
character; "extended" (default) or "simple". "extended" yields a plotted tree with observed data and/or fitted means in the terminal nodes; "simple" yields a plotted tree with fixed and/or random effects coefficients presented in the terminal nodes.

observed
logical. Should observed datapoint be plotted?

fitmean
logical. Should the fitted mean function be visualized? Only used when predictor variables for the node-specific (G)LMs are specified.

fit.ranef
character; "fixed" (default) or "varying". Only used when fitmean = TRUE. If "constant", fitted values will be calculated, with random effects fixed at their population mean (0). If "varying", fitted values will also be based on the fitted random effects.

fit.fixef
character, "constant" (default) or "varying". Only used when fitmean = TRUE. If "constant", fitted values will be calculated, keeping the values of the other predictors in the (G)LM constant at their sample mean or majority value. If "varying", fitted values will also be based on the fitted effects of the remaining predictor variables.

tp_args
list of arguments to be passed to panel generating function node_glmertree. See arguments node_bivplot in panelfunctions.

drop_terminal
logical. Should all terminal nodes be plotted at the bottom?

...
Additional arguments to be passed to plot.party(). See party-plot documentation for details.

Details

The code is still under development and might change in future versions.

Value

The function returns a list with the following objects:

tree
The final lmtree/glmtree.

lmer
The final lmer random-effects model.

ranef
The corresponding random effects of lmer.

varcorr
The corresponding VarCorr(lmer).

variance
The corresponding attr(VarCorr(lmer), "sc")^2.

data
The dataset specified with the data argument including added auxiliary variables .ranef and .tree from the last iteration.

loglik
The log-likelihood value of the last iteration.

iterations
The number of iterations used to estimate the lmertree.

maxit
The maximum number of iterations specified with the maxit argument.

ranefstart
The random effects used as an offset, as specified with the ranefstart argument.

formula
The formula as specified with the formula argument.

randomformula
The formula as specified with the randomformula argument.

abstol
The prespecified value for the change in log-likelihood to evaluate convergence, as specified with the abstol argument.

mob.control
A list containing control parameters passed to lmtree(), as specified with ....

lmer.control
A list containing control parameters passed to lmer(), as specified in the lmer.control argument.

joint
Whether the fixed effects from the tree were (re-)estimated jointly along with the random effects, specified with the joint argument.

References

Fokkema M, Smits N, Zeileis A, Hothorn T, Kelderman H (2018). “Detecting Treatment-Subgroup Interactions in Clustered Data with Generalized Linear Mixed-Effects Model Trees”. Behavior Research Methods, 50(5), 2016-2034. https://doi.org/10.3758/s13428-017-0971-x

See Also

lmertree, glmertree, party-plot.

Examples

## load artificial example data
data("DepressionDemo", package = "glmertree")

## fit normal linear regression LMM tree for continuous outcome
lt <- lmertree(depression ~ treatment + age | cluster | anxiety + duration,
  data = DepressionDemo)
plot.lmertree2(lt)
plot.lmertree2(lt, type = "simple")
plot.lmertree2(lt, which = "tree", fit.ranef = "constant", fit.fixef = "varying")
plot.lmertree2(lt, which = "tree", fit.ranef = "varying", fit.fixef = "constant")
plot.lmertree2(lt, which = "tree", fit.ranef = "varying", fit.fixef = "varying")

## fit logistic regression GLMM tree for binary outcome
gt <- glmertree(depression_bin ~ treatment | cluster | 
  age + anxiety + duration, data = DepressionDemo)
plot.glmertree2(gt)
plot.glmertree2(gt, type = "simple")
plot.glmertree2(gt, which = "tree", fit.ranef = "constant", fit.fixef = "varying")
plot.glmertree2(gt, which = "tree", fit.ranef = "varying", fit.fixef = "constant")
plot.glmertree2(gt, which = "tree", fit.ranef = "varying", fit.fixef = "varying")

[Package glmertree version 0.1-3 Index]